new hook to override area parameters
In the area of conversations with Amadeus, english Steam version
of SG0 moved the position of the rotating ring in the end down,
presumably to compensate for the expansion of the background rectangle.
CoZ patch fixes the background image, and the rotating ring feels
somewhat out of place. This code allows to specify the parameters
including the position of the rotating ring (and also other related
parameters just in case).