user-controlled option for disabling controllers
By default, the game processes everything that looks like a controller,
even if it was not configured in the launcher. This behaviour can be
good or bad depending on the situation. It is good for users
who actually have a gamepad and rush to the game without checking
all the settings beforehand. It is bad for users whose systems
have some strange-behaving devices, especially for users who are not aware
of these devices.

88ef10 unconditionally disables controllers that are not explicitly
configured in the launcher. This fixes the game for the second category
of users at the cost of confusing the first category.

Make this behaviour to be controlled by an user option.

I think the default should be false so that it just works for most users
(the second category seems to be much smaller than the first one)
while still offering a solution for affected users.
Plus, it would stop complaints like "my gamepad worked,
I installed the patch, now it doesn't". I don't insist, however.