experimental feature: partial video redirection
Sometimes it is useful to override only a part of existing video.
One concrete example is to translate SG0 intro and true-ending videos:
the former one starts and the latter one ends with text phrases
with video effects that could be translated by someone
with video editing skills. The global video redirection surely works,
but the patch size becomes quite large. Subtitles are a viable alternative,
but they add to original rather than override it, and it is not easy
or even possible to replicate video effects.

Config example:
"videoPartialRedirection": {
  "name.bk2": [
    {"name": "redirected.bk2", "startFrame": 1234, "frameCount": 42, "preload": true}
  ]
}
Some notes:
* I'm not even trying to do anything with audio. Gapless playback
  is a pain. A possible interaction with global audio redirection
  does not exactly make it easier.
  Any audio in the partial video is ignored.
* As a consequence, while the partial video is active,
  disk I/O goes on for both videos, BINK still needs to read
  the original one for audio. For short parts, it can be mitigated
  with "preload":true, this flag loads the entire file when
  the original video is opened.
* The code tries to avoid decoding the original video during the redirection,
  but it is not always possible. If we need to switch back
  and there are no keyframes in the original video, the original video
  must still be decoded even while the partial video is active.
* It seems to work in basic cases, but there are probably some corner cases
  that I have not thought of.