fix pixel length of speakers' names
When Steam versions of S;G 0 and C;C (and probably some others, but not S;G)
draw a speaker's name at the bottom of a dialog, there is
a little discrepancy between how the length in pixels is computed
to define the bounding area, and the actual layout.
Namely, since these use a smaller font, the scaling results
in fractional sizes for every character. The actual layout just drops
the fractional part. The calculation of the total length keeps 7 bits
after the decimal point, i.e. uses floor(<ideal size> * 0x80).
As a result, the name is not actually centered in the bounding box
allocated for it.

Almost all names are relatively short, so the effect is just
a couple of pixels, only hardcore perfectionists can notice this.
However, occasionally longer names appear
("Girl with the short haircut"), the discrepancy is noticeable for them.

The patch just swaps two instructions, making the formula
floor(<ideal size> * 0x80) -> floor(<ideal size>) * 0x80.

Signature for the current Steam version of SG0 and CC:
C1 E0 07 F7 F1 8B ?? ?? ?? ?? ?? BA 02 80 00 00.